Changes in sqlite struct and in BtreeCursor.
Updates in tarantool API in sqlite struct: new functions for BtreeCursors emulation.
Updates in sqlite BtreeCursor: new member of struct BtCursor - flag if this
object is cursor on tarantool or on sqlite Btree and pointer on TarantoolCursor
object if flag is set. Also cursors interface was modified: now if flag of
BtCuror is set then call will be passed to functions from tarantool_api in
struct sqlite3 object.

Signed-off-by: Vladislav Shpilevoy <v.shpilevoy@corp.mail.ru>